Add ability to configure EUI-64 addresses and to enable auto configuration.
The first mode allows the administrator to configure an EUI-64 format
IPv6 address without having to enter the low-order 64-bits.  The
administrator sets only the high-order 64-bits, and the system
generates the low-order bits based on the MAC address.

The second mode enables using stateless address auto-configuration
(SLAAC) to acquire an IPv6 address.  But since this is a "host"
feature, it is only enabled if IPv6 forwarding is disabled.